Reduce size of some files by extracting tests
