+++ src/common/tests/mod.rs
@@ -1,2 +1,4 @@
 mod placement_tests;
 mod scrolling_strips_tests;
+mod sizing_tests;
+mod workspace_tests;
